Review fixes

+++ contracts/ColonyStorage.sol
@@ -31,41 +31,41 @@ contract ColonyStorage is DSAuth {
 
   address resolver;
 
-  mapping (uint => Transaction) transactions;
+  mapping (uint256 => Transaction) transactions;
   // Mapping function signature to 2 task roles whose approval is needed to execute
   mapping (bytes4 => uint8[2]) reviewers;
   // Maps transactions to roles and whether they've confirmed the transaction
-  mapping (uint => mapping (uint => bool)) confirmations;
-  uint transactionCount;
+  mapping (uint256 => mapping (uint256 => bool)) confirmations;
+  uint256 transactionCount;
 
   struct Transaction {
     bytes data;
-    uint value;
+    uint256 value;
     bool executed;
   }
 
   address colonyNetworkAddress;
   ERC20Extended token;
-  mapping (uint => Task) tasks;
+  mapping (uint256 => Task) tasks;
 
   // Pots can be tied to tasks or to (in the future) domains, so giving them their own mapping.
   // Pot 1  can be thought of as the pot belonging to the colony itself that hasn't been assigned
   // to anything yet, but has had some siphoned off in to the reward pot.
   // Pot 0 is the pot containing funds that can be paid to holders of colony tokens in the future.
-  mapping (uint => Pot) pots;
+  mapping (uint256 => Pot) pots;
 
   // This keeps track of how much of the colony's funds that it owns have been moved into pots other than pot 0,
   // which (by definition) have also had the reward amount siphoned off and put in to pot 0.
   // TODO: This needs to be decremented whenever a payout occurs and the colony loses control of the funds.
-  mapping (address => uint) nonRewardPotsTotal;
+  mapping (address => uint256) nonRewardPotsTotal;
 
-  mapping (uint => RatingSecrets) public taskWorkRatings;
+  mapping (uint256 => RatingSecrets) public taskWorkRatings;
 
-  mapping (uint => Domain) public domains;
+  mapping (uint256 => Domain) public domains;
 
-  uint taskCount;
-  uint potCount;
-  uint domainCount;
+  uint256 taskCount;
+  uint256 potCount;
+  uint256 domainCount;
 
   struct Task {
     bytes32 specificationHash;
@@ -80,11 +80,11 @@ contract ColonyStorage is DSAuth {
     uint256[] skills;
 
     // TODO switch this mapping to a uint8 when all role instances are uint8-s specifically ColonyFunding source
-    mapping (uint => Role) roles; 
+    mapping (uint256 => Role) roles; 
     // Maps a token to the sum of all payouts of it for this task
-    mapping (address => uint) totalPayouts;
+    mapping (address => uint256) totalPayouts;
     // Maps task role ids (0,1,2..) to a token amount to be paid on task completion
-    mapping (uint => mapping (address => uint)) payouts;
+    mapping (uint256 => mapping (address => uint256)) payouts;
   }
 
   struct Role {
@@ -103,13 +103,13 @@ contract ColonyStorage is DSAuth {
   }
 
   struct Pot {
-    mapping (address => uint) balance;
-    uint taskId;
+    mapping (address => uint256) balance;
+    uint256 taskId;
   }
 
   struct Domain {
-    uint skillId;
-    uint potId;
+    uint256 skillId;
+    uint256 potId;
   }
 
   modifier taskExists(uint256 _id) {
